Bisection
In mathematics, the bisection method is a simple methodology for locating numerical solutions to equations containing a single unknown. Among all numerical approaches for solving the transcendental equation, the bisection method is the easiest. We will explore the bisection approach with solved issues in-depth in this article.
Bisection is a technique for determining the roots of a polynomial equation. It splits and separates the interval containing the root of the equation. The intermediate theorem for continuous functions serves as the foundation for this strategy. It operates by closing the gap between the positive and negative intervals until the proper solution is found. By averaging the positive and negative intervals, this approach closes the gap. It is a straightforward procedure that is quite slow. Bisection is also referred to as interval halving, root-finding, binary search, or dichotomy approach.